Типы данных для призовых переменных. Отображение терминалов, для которых нет счетчиков в БД. Корректировка отображения общего входа и баланса

+++ Server/Models/PrizeFundPageModel.cs
@@ -8,8 +8,8 @@ namespace Server.Models
         public Guid GroupId { get; set; }
         public string GroupName { get; set; }
         public bool Status { get; set; }
-        public uint ValuePrize { get; set; }
-        public uint Fund { get; set; }
+        public ushort ValuePrize { get; set; }
+        public double Fund { get; set; }
     }
     
     public class PrizeFundPageModel
